requirements.txt
httpx~=0.28.1
fastapi~=0.115.7
pydantic~=2.10.6
SQLAlchemy~=2.0.37
alembic~=1.14.1
aiogram~=3.17.0
python-dotenv~=1.0.1